Implementare entitati si teste pentru daily forcasting

+++ src/main/java/com/arobs/weather/snapshot/Rain.java
@@ -10,15 +10,15 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 public class Rain {
 
     @JsonProperty("3h")
-    private Integer _3h;
+    private Double _3h;
 
     @JsonProperty("3h")
-    public Integer get3h() {
+    public Double get3h() {
         return _3h;
     }
 
     @JsonProperty("3h")
-    public void set3h(Integer _3h) {
+    public void set3h(Double _3h) {
         this._3h = _3h;
     }
 }
